inference-services/python-onnx-service/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
gliner>=0.1.13
torch>=2.0.0
transformers>=4.35.0
